requirements.txt
langchain
langchain-core
langchain-huggingface
huggingface-hub
python-dotenv
numpy
scikit-learn